+++ src/command_modules/azure-cli-redis/azure/cli/command_modules/redis/tests/latest/test_redis_scenario.py
@@ -13,19 +13,22 @@ class RedisCacheTests(ScenarioTest):
 
         self.kwargs = {
             'rg': resource_group,
-            'loc': 'WestUS',
+            'loc': 'WestUS2',
             'name': self.create_random_name(prefix='redis', length=24),
             'sku': 'premium',
             'sku2': 'basic',
             'size': 'P1',
             'name2': self.create_random_name(prefix='redis', length=24),
             'size2': 'C0',
+			'tls_version': '1.2',
+			'tags': "{\\\"test\\\":\\\"tryingzones\\\"}",
+			'zones': "{\\\"1\\\",\\\"2\\\"}",
             'settings': "{\\\"hello\\\":1}",
             'schedule_entries': "[{\\\"dayOfWeek\\\":\\\"Monday\\\",\\\"startHourUtc\\\":\\\"00\\\",\\\"maintenanceWindow\\\":\\\"PT5H\\\"}]",
             'schedule_entries_update': "[{\\\"dayOfWeek\\\":\\\"Tuesday\\\",\\\"startHourUtc\\\":\\\"01\\\",\\\"maintenanceWindow\\\":\\\"PT10H\\\"}]"
         }
 
-        self.cmd('az redis create -n {name} -g {rg} -l {loc} --sku {sku} --vm-size {size}')
+        self.cmd('az redis create -n {name} -g {rg} -l {loc} --sku {sku} --vm-size {size} --minimum-tls-version {tls_version} --tags {tags} --zones {zones}')
 
         self.cmd('az redis create -n {name2} -g {rg} -l {loc} --sku {sku2} --vm-size {size2} --tenant-settings {settings}')
         self.cmd('az redis show -n {name} -g {rg}', checks=[